create all global indices via eventindex

+++ coffea/nanoevents/transforms.py
@@ -339,5 +339,5 @@ def item(stack):
 
 def eventindex(stack):
     out = stack.pop()
-    out, _ = awkward.broadcast_arrays(numpy.arange(len(out)), out)
+    out, _ = awkward.broadcast_arrays(numpy.arange(len(out), dtype=numpy.int64), out)
     stack.append(out)
